# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Semantic Resume Matcher",
    "version": "0.1.0"
  },
  "paths": {
...
    "/candidate/{candidate_id}": {
      "get": {
        "tags": [
          "results"
        ],
        "summary": "Get Candidate Status",
        "description": "Retrieve the pipeline execution status of a candidate.\n\nAllows clients to poll the processing state (e.g. PENDING, PROCESSING, COMPLETE, FAILED)\nof a candidate's resume upload.",
        "operationId": "get_candidate_status_candidate__candidate_id__get",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/candidate/{candidate_id}/matches": {
      "get": {
        "tags": [
          "results"
        ],
        "summary": "Get Match Results",
        "description": "Retrieve semantic and deterministic match results for a candidate.\n\nReturns job matches sorted by overall confidence and semantic similarity score.\nIncludes fit reasoning, strengths, and missing candidate skill gaps.",
        "operationId": "get_match_results_candidate__candidate_id__matches_get",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResultsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CandidateStatusResponse": {
        "properties": {
          "candidate_id": {
            "type": "string",
            "format": "uuid",
            "title": "Candidate Id"
          },
          "pipeline_status": {
            "type": "string",
            "title": "Pipeline Status"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "candidate_id",
          "pipeline_status",
          "status"
        ],
        "title": "CandidateStatusResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MatchJobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "remote_ok": {
            "type": "boolean",
            "title": "Remote Ok"
          },
          "min_salary": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Salary"
          },
          "max_salary": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Salary"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "title",
          "company",
          "remote_ok"
        ],
        "title": "MatchJobResponse"
      },
      "MatchResultItem": {
        "properties": {
          "match_id": {
            "type": "string",
            "format": "uuid",
            "title": "Match Id"
          },
          "job": {
            "$ref": "#/components/schemas/MatchJobResponse"
          },
          "vector_score": {
            "type": "number",
            "title": "Vector Score"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "match_category": {
            "type": "string",
            "title": "Match Category"
          },
          "reasoning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning"
          },
          "skill_gaps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skill Gaps"
          },
          "standout_strengths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Standout Strengths"
          }
        },
        "type": "object",
        "required": [
          "match_id",
          "job",
          "vector_score",
          "match_category",
          "skill_gaps",
          "standout_strengths"
        ],
        "title": "MatchResultItem"
      },
      "MatchResultsResponse": {
        "properties": {
          "candidate_id": {
            "type": "string",
            "format": "uuid",
            "title": "Candidate Id"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/MatchResultItem"
            },
            "type": "array",
            "title": "Matches"
          }
        },
        "type": "object",
        "required": [
          "candidate_id",
          "matches"
        ],
        "title": "MatchResultsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}